+++ src/main/java/oshi/jna/platform/linux/package-info.java
@@ -12,6 +12,6 @@
  * https://github.com/dblock/oshi/graphs/contributors
  */
 /**
- * Provides extensions of JNA libraries for Linux
+ * Provides extensions of JNA libraries for Linux.  These classes should be considered non-API as they may be removed if/when their code is incorporated into the JNA project.
  */
 package oshi.jna.platform.linux;