+++ src/main/java/oshi/jna/platform/mac/CoreFoundation.java
@@ -23,7 +23,9 @@ import com.sun.jna.Pointer;
 import com.sun.jna.PointerType;
 
 /**
- * CoreFoundation framework for power supply stats
+ * CoreFoundation framework for power supply stats. This class should be
+ * considered non-API as it may be removed if/when its code is incorporated into
+ * the JNA project.
  * 
  * @author widdis[at]gmail[dot]com
  */