+++ src/main/java/oshi/jna/platform/windows/PowrProf.java
@@ -26,7 +26,8 @@ import com.sun.jna.Pointer;
 import com.sun.jna.Structure;
 
 /**
- * Power profile stats
+ * Power profile stats. This class should be considered non-API as it may be
+ * removed if/when its code is incorporated into the JNA project.
  * 
  * @author widdis[at]gmail[dot]com
  */