+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 2.0 (in-progress)
 ================
 * [#101](https://github.com/dblock/oshi/pull/101): Refactored package structure for consistency - [@dbwiddis](https://github.com/dbwiddis).
+* [#103](https://github.com/dblock/oshi/pull/103): Switched CentralProcessor to a single object for all processors - [@dbwiddis](https://github.com/dbwiddis).
 * Your contribution here.
 
 1.5.2 (11/23/2015)